+++ writer.go
@@ -2,72 +2,62 @@ package tower
 
 import (
 	"io"
-	"strings"
 )
 
-type Writer interface {
+type LineWriter interface {
 	io.Writer
 	io.StringWriter
-
-	// Returns the accumulated values.
-	String() string
-}
-
-type LineWriter interface {
-	Writer
-	// Writes a predetermined separator to the writer.
+	// WriteSeparator Writes a predetermined separator to the writer.
 	WriteSeparator()
-	// Writes a predetermined prefix to the writer.
+	// WritePrefix Writes a predetermined prefix to the writer.
 	WritePrefix()
-	// Writes a predetermined suffix to the writer.
+	// WriteSuffix Writes a predetermined suffix to the writer.
 	WriteSuffix()
-	// Writes Indentation characters.
+	// WriteIndent Writes Indentation characters.
 	WriteIndent()
+	// GetSeparator Returns the pre-determined separator.
 	GetSeparator() string
+	// GetPrefix Returns the pre-determined prefix.
 	GetPrefix() string
+	// GetSuffix Returns the pre-determined suffix.
 	GetSuffix() string
+	// GetIndentation Returns the pre-determined indentation.
 	GetIndentation() string
 }
 
 type LineWriterBuilder struct {
-	writer    Writer
+	writer    io.Writer
 	indent    string
 	separator string
 	prefix    string
 	suffix    string
 }
 
-// Sets the Writer target.
-func (builder *LineWriterBuilder) Writer(w Writer) *LineWriterBuilder {
-	builder.writer = w
-	return builder
-}
-
-// Sets the Indentation.
+// Indent Sets the Indentation.
 func (builder *LineWriterBuilder) Indent(s string) *LineWriterBuilder {
 	builder.indent = s
 	return builder
 }
 
-// Sets the Linebreak character(s).
+// Separator Sets the Linebreak character(s).
 func (builder *LineWriterBuilder) Separator(s string) *LineWriterBuilder {
 	builder.separator = s
 	return builder
 }
 
-// Sets the Prefix.
+// Prefix Sets the Prefix.
 func (builder *LineWriterBuilder) Prefix(s string) *LineWriterBuilder {
 	builder.prefix = s
 	return builder
 }
 
-// Sets the Suffix.
+// Suffix Sets the Suffix.
 func (builder *LineWriterBuilder) Suffix(s string) *LineWriterBuilder {
 	builder.suffix = s
 	return builder
 }
 
-// Turn this writer into proper LineWriter.
+// Build Turn this writer into proper LineWriter.
 func (builder *LineWriterBuilder) Build() LineWriter {
 	return &lineWriter{
 		Writer:    builder.writer,
@@ -78,29 +68,33 @@ func (builder *LineWriterBuilder) Build() LineWriter {
 	}
 }
 
-// Creates a new LineWriterBuilder. You have to call .Build() to actually use LineWriter.
-func NewLineWriterBuilder() *LineWriterBuilder {
+// NewLineWriter Creates a new LineWriterBuilder. You have to call .Build() to actually use LineWriter.
+func NewLineWriter(writer io.Writer) *LineWriterBuilder {
 	return &LineWriterBuilder{
-		writer: &strings.Builder{},
+		writer: writer,
 	}
 }
 
 var _ LineWriter = (*lineWriter)(nil)
 
 type lineWriter struct {
-	Writer
+	io.Writer
 	separator string
 	prefix    string
 	suffix    string
 	indent    string
 }
 
-// Writes Indentation characters.
+func (l *lineWriter) WriteString(s string) (n int, err error) {
+	return io.WriteString(l.Writer, s)
+}
+
+// WriteIndent Writes Indentation characters.
 func (l *lineWriter) WriteIndent() {
 	if lw, ok := l.Writer.(LineWriter); ok {
 		lw.WriteIndent()
 	}
-	_, _ = l.Writer.WriteString(l.indent)
+	_, _ = io.WriteString(l.Writer, l.indent)
 }
 
 func (l lineWriter) GetIndentation() string {
@@ -114,21 +108,21 @@ func (l *lineWriter) WriteSeparator() {
 	if lw, ok := l.Writer.(LineWriter); ok {
 		lw.WriteSeparator()
 	}
-	_, _ = l.WriteString(l.separator)
+	_, _ = io.WriteString(l.Writer, l.separator)
 }
 
 func (l *lineWriter) WritePrefix() {
 	if lw, ok := l.Writer.(LineWriter); ok {
 		lw.WritePrefix()
 	}
-	_, _ = l.WriteString(l.prefix)
+	_, _ = io.WriteString(l.Writer, l.prefix)
 }
 
 func (l *lineWriter) WriteSuffix() {
 	if lw, ok := l.Writer.(LineWriter); ok {
 		lw.WriteSuffix()
 	}
-	_, _ = l.WriteString(l.suffix)
+	_, _ = io.WriteString(l.Writer, l.suffix)
 }
 
 func (l lineWriter) GetSeparator() string {
@@ -153,26 +147,26 @@ func (l lineWriter) GetSuffix() string {
 }
 
 type Display interface {
-	// Display returns a human readable and rich with information for the implementer.
+	// Display returns a human-readable and rich with information for the implementer.
 	Display() string
 }
 
 type DisplayWriter interface {
-	// Writes the Display() string to the writer instead of being allocated as value.
+	// WriteDisplay Writes the Display() string to the writer instead of being allocated as value.
 	WriteDisplay(w LineWriter)
 }
 
 type ErrorWriter interface {
-	// Writes the error.Error to the writer instead of being allocated as value.
+	// WriteError Writes the error.Error to the writer instead of being allocated as value.
 	WriteError(w LineWriter)
 }
 
 type Summary interface {
-	// Returns a short summary of the implementer.
+	// Summary Returns a short summary of the implementer.
 	Summary() string
 }
 
 type SummaryWriter interface {
-	// Writes the Summary() string to the writer instead of being allocated as value.
+	// WriteSummary Writes the Summary() string to the writer instead of being allocated as value.
 	WriteSummary(w LineWriter)
 }